requirements.txt
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Navigation==0.2.0
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-decouple==3.4
requests==2.25.1
urllib3==1.26.3
Werkzeug==1.0.1
WTForms==2.3.3
